feat: add frontend views and file management functionality, enhance user response structure, and update configuration guide
