+++ app/__init__.py
@@ -1,0 +1,3 @@
+"""
+Document processor application package.
+"""
