+++ app/main.py
@@ -15,9 +15,9 @@ from app.database import init_db
 from app.config import settings
 from app.utils.config_validator import check_all_configs
 
-# Import both the traditional and new routers
-from app.frontend import router as frontend_router_original
-from app.api import router as api_router  # Updated import for the refactored API router
+# Import the routers - now using views directly instead of frontend
+from app.views import router as frontend_router
+from app.api import router as api_router
 from app.auth import router as auth_router
 
 # Load configuration from .env for the session key
@@ -95,8 +95,8 @@ def test_500():
     raise RuntimeError("Testing forced 500 error!")
 
 # Include the routers
-app.include_router(frontend_router_original)
+app.include_router(frontend_router)
 app.include_router(auth_router)
-app.include_router(api_router, prefix="/api")  # Use the refactored API router
+app.include_router(api_router, prefix="/api")
